requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
twilio>=8.10.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
